rename language variables
